rename velocities to something more specific
